JUPYTER: Phase 1 — graduate scout_xnat_auth to a real package
