+++ xnat/scout-xnat-auth/scout_xnat_auth/errors.py
@@ -1,0 +1,3 @@
+class ScoutXnatAuthError(RuntimeError):
+    """Raised when the bearer-token bootstrap fails (env vars missing,
+    Hub API rejected, XNAT rejected the token, etc.)."""
